Model Feature working. Front-end fixed, and Billing AJAX working, also MRN setting.
Some classes are created, while some are been refactored, like the QueryBuilder(Yes, it's a silly name tho).
Clinicians can now query Patients with their MRN, and view their Billing Datas.
Lots of abstractive classes are in place, the rest of the API should be easy to build now.
Probably, no COMMIT tomorrow. It is Sunday!